Change from int to double

+++ ConCurrency/Cashier.hpp
@@ -12,7 +12,7 @@ private:
 public:
 	Cashier(std::shared_ptr<ExchangeCurrency> exchange, std::shared_ptr<CashRegister> cashRegister);
 
-	std::pair <int, double> exchangeCurrencyforCashier(int amount, std::string CurrencyCode);
+	std::pair <double, double> exchangeCurrencyforCashier(int amount, std::string CurrencyCode);
 
 	void displayOpeningBalance();
 	void displayCurrencyRateTable();
